debug: Add comprehensive debug logging to trace interpolation execution
- Add debug logging throughout interpolation pipeline to identify hanging issues
- Fix parameter passing in convert.py to include all interpolation config parameters
- Add timing logs for get_common_mass_axis() which is likely causing hangs
- Add bounds detection debug logging
- Add streaming engine and dask pipeline debug tracing

+++ msiconvert/core/base_reader.py
@@ -242,6 +242,6 @@ class BaseMSIReader(ABC):
             
             # Fill buffer with spectrum data
             buffer.coords = coords
-            buffer.fill(mzs, intensities.astype(np.float32))
+            buffer.fill(mzs, intensities.astype(np.float64))
             
             yield buffer
